Add basic echoserver and echoclient.
I added the basic echo server and client, as well as the lsplog, lspnet, and
lsp12 libraries for use in the project. We'll use the lsp12 library to connect
to servers since it has the heartbeat sensor already built in.

The next step is to write a basic test framework for this set up. That way, we
can modify it easily as we add components to the system.