Paxos
I finished implementing the Paxos protocol and the testing framework. The
makefile contains recipes for creating a pool of servers and a client. It also
contains a recipe for killing a single server during testing.

I'll make the echostore server an RPC server and integrate it into the existing
framework.